+++ vector.h
@@ -1,11 +1,11 @@
 struct vector {
 	/* coordinates */
-	double x, y, z;
+	float x, y, z;
 };
 
-struct vector makevec(double, double, double);
+struct vector makevec(float, float, float);
 struct vector add(struct vector, struct vector);
 struct vector sub(struct vector, struct vector);
-struct vector scale(double, struct vector);
+struct vector scale(float, struct vector);
 struct vector norm(struct vector a);
-double dot(struct vector, struct vector);
+float dot(struct vector, struct vector);
